refactor commands to subcommands

+++ src/controllers.ts
@@ -1,3 +1,4 @@
+import { ApplicationCommandOptionType } from 'discord-api-types/payloads/v10';
 import { InteractionResponseType, InteractionType } from 'discord-interactions';
 import { NextFunction, Request, Response } from 'express';
 import FormData from 'form-data';
@@ -5,6 +6,7 @@ import { createReadStream, unlink } from 'fs';
 import { basename } from 'path';
 import cataas from './cataas';
 import config from './config';
+import discord from './discord';
 import { writeReadableToFile } from './file';
 import logger from './logger';
 
@@ -37,15 +39,12 @@ export const processInteraction = method(
 
     if (type === InteractionType.APPLICATION_COMMAND) {
       const { name, options } = data;
-      const commandOptions =
-        options?.reduce((accum: any, option: any) => {
-          accum[option.name] = option.value;
-          return accum;
-        }, {}) || {};
+      const commandOptions = discord.readCommandOptions(options);
       const serializedOptions = Object.keys(commandOptions)
         .map((optName) => {
           let value = commandOptions[optName];
           if (typeof value === 'string') value = `"${value}"`;
+          if (typeof value === 'object') value = `${JSON.stringify(value)}`;
           return `${optName}: ${value}`;
         })
         .join(', ');
@@ -56,35 +55,44 @@ export const processInteraction = method(
 
       switch (name) {
         case 'cat':
-          const cataasRes = await cataas.getCat(commandOptions);
-          const fp = await writeReadableToFile(
-            cataasRes.body,
-            config.dirTmpImg,
-            { mime: cataasRes.headers.get('content-type') },
-          );
-
-          const formData = new FormData();
-          formData.append(
-            'payload_json',
-            JSON.stringify({
-              type: InteractionResponseType.CHANNEL_MESSAGE_WITH_SOURCE,
-              content: 'Received',
-              attachments: [{ id: 0, name: basename(fp) }],
-            }),
-            { contentType: 'application/json' },
-          );
-          formData.append('files[0]', createReadStream(fp), basename(fp));
-          res.set(formData.getHeaders());
-          res.once('close', () => {
-            unlink(fp, (err) => {
-              if (err) {
-                logger.warn(`Unable to delete file ${fp}`, err);
-              } else {
-                logger.debug(`Deleted file ${fp}`);
-              }
-            });
+          res.json({
+            type: InteractionResponseType.DEFERRED_CHANNEL_MESSAGE_WITH_SOURCE,
           });
-          return formData.pipe(res);
+          // deferred response
+          let fp: string | null = null;
+          try {
+            const cataasRes = await cataas.getCat(commandOptions);
+            fp = await writeReadableToFile(cataasRes.body, config.dirTmpImg, {
+              mime: cataasRes.headers.get('content-type'),
+            });
+
+            const formData = new FormData();
+            formData.append(
+              'payload_json',
+              JSON.stringify({
+                type: InteractionResponseType.CHANNEL_MESSAGE_WITH_SOURCE,
+                attachments: [{ id: 0, name: basename(fp) }],
+              }),
+              { contentType: 'application/json' },
+            );
+            formData.append('files[0]', createReadStream(fp), basename(fp));
+            await discord.interactionFollowup(req.body.token, {
+              body: formData,
+            });
+          } catch (err: any) {
+            logger.error(`Unable to reply to an interaction`, err);
+          } finally {
+            if (fp) {
+              unlink(fp, (err) => {
+                if (err) {
+                  logger.warn(`Unable to delete file ${fp}`, err);
+                } else {
+                  logger.debug(`Deleted file ${fp}`);
+                }
+              });
+            }
+          }
+          return;
         default:
           return res.status(400).json({ message: 'Unknown command' });
       }
